MSEARCH-794: refactor service classes and tests
