+++ src/main/java/org/folio/search/controller/IndexManagementController.java
@@ -16,7 +16,7 @@ import org.folio.search.rest.resource.IndexManagementApi;
 import org.folio.search.service.IndexService;
 import org.folio.search.service.ResourceService;
 import org.folio.search.service.reindex.ReindexService;
-import org.folio.search.service.reindex.ReindexUploadRangeIndexService;
+import org.folio.search.service.reindex.ReindexStatusService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -33,9 +33,9 @@ import org.springframework.web.bind.annotation.RestController;
 public class IndexManagementController implements IndexManagementApi {
 
   private final IndexService indexService;
-  private final ReindexUploadRangeIndexService reindexRangeService;
   private final ResourceService resourceService;
   private final ReindexService reindexService;
+  private final ReindexStatusService reindexStatusService;
 
   @Override
   public ResponseEntity<FolioCreateIndexResponse> createIndices(String tenantId, CreateIndexRequest request) {
@@ -74,6 +74,6 @@ public class IndexManagementController implements IndexManagementApi {
 
   @Override
   public ResponseEntity<List<ReindexStatusItem>> getReindexStatus(String tenantId) {
-    return ResponseEntity.ok(reindexRangeService.getReindexStatuses(tenantId));
+    return ResponseEntity.ok(reindexStatusService.getReindexStatuses(tenantId));
   }
 }
